[Parallel IO]
int MPI_File_read(
  MPI_File mpi_fh,
  void *buf,
  int count,
  MPI_Datatype datatype,
  MPI_Status *status
);

In the above interface, 'datatype' describes the access pattern of the output user buffer - 'buf'
However, this data type is not imposed upon the file itself. The number of bytes to be read from the file = the number of bytes available in the 'usable' portion of the userbuffer is:
num_bytes_user_buff = ('usable' bytes in 1 element of type 'datatype' ) x (count)

This type - 'datatype' is not imposed upon the fileview while reading. The number of file etypes to be read form the file is:
num_etypes = (num_bytes_user_buff) / (size of one etype in bytes)